feat(v2): new type-safe API introduced
this commit introduces the first needed work for a new API for papercut.
This new API does not uses classes, is completely type safe, uses more
common terminology for dealing with HTML nodes and elements, and also
introduces a new `strict` option to the scrape runner which allows users
to turn off the `suppress` function when executing the scraping
selectors.